Rename upstream CircleCI configuration
We're not interested in triggering the extremely slow upstream CircleCI
configuration, since it takes forever to run and burns a lot of cash.

This commit renames it so that we can add our own configuration without
conflicts.

First cut of RocksDB+circlci+cmake on Linux, Mac(x86_64,arm64), Windows

etl3
====
1. Adding sm-rocksdb:sm-rocksdbjni = 7.7.3-SNAPSHOT to version.props & write-locks.
2. Adding implementation("sm-rocksdb:sm-rocksdbjni") to streaming-code/build.gradle.kts

version
=======
* if commit is tagged with `sm-v7.7.3`, a 7.7.3 version release is pushed to jfrog.
* Else the version is retrieved from version.h, and -SNAPSHOT version is pushed to jfrog.

source
======
* Any addition/removal of java src or test file will have to be updated in java/CMakeLists.txt
* Any addition/removal of test class will have to be updated in class SMRocksJunitRunner

circleci
========
 { build-linux, build-macos, build-windows} contribute jar, test-jar, and jni libraries.
 { make-jar } merges the jni librariesin the jar.
 { test-linux, test-macos, test-windows} run the test-jar on repsective platforms,
 { upload } uploads the merged-jar to artifactory.

pending
=======
1. Add other compressions.

Upload artifacts so that we can test them on M1 macs

Ensure that Linux and Windows builds are portable

Discovered when I tried to run the output JAR on my workstation and
it failed when it hit an AVX-512 instruction.

Fix warning in Windows build

Trying newer version of liburing, and misc

Make JNI test header dependency explicit

A few of the unit tests depend on JNI headers, so make sure that we
generate the JNI headers explicitly in the build process.

Fix transitive dependencies

Adding names to circlci, indenting java file, and a typo

Rename "rocksdbjni-shared" target to "rocksdbjni"

The Java code expects the default JNI library to be named "librocksdbjni.so"
or the equivalent, not "librocksdbjni-shared.so". This commit renames the build
target so that it's easy to run the build output from the command line or
IntelliJ without renaming the shared library.

(Technically, this removes the ability to build a static JNI library. In
practice, I've never heard of anyone using a static JNI library; while it's
technically possible, it requires rebuilding and relinking the JVM.)

Using junit5 console launcher for automatic test detection

Put all tests classes in JAVA_TEST_CLASSES

This ensures that all of the test classes are present so that they can
be discovered by the JUnit5 runner.

Hook up CTest for java tests

This commit enables support for running the rocksdbjni tests using CTest.
This is a bit janky (we need to download the junit5 runner from maven),
but it removes some redundant package-munging and makes it possible to
run the tests without hand-crafting the correct command-line invocation.