Implement automatic test discovery
The list of tests in SMRocksJunitRunner was copied manually from the Makefile,
so it will not be updated when new tests are added upstream. Even if new tests
are included in the test JAR, they still will not be picked up. This could lead
us to inadvertently break new functionality with our changes.

This commit adds a small test discovery implementation to the test runner. The
test discovery mechanism is used unless a list of test classes are specified
on the command line. The discovery code works for both JARs and directories of
class files, so it can (theoretically) be run without rebuilding the JAR.