Rename "rocksdbjni-shared" target to "rocksdbjni"
The Java code expects the default JNI library to be named "librocksdbjni.so"
or the equivalent, not "librocksdbjni-shared.so". This commit renames the build
target so that it's easy to run the build output from the command line or
IntelliJ without renaming the shared library.

(Technically, this removes the ability to build a static JNI library. In
practice, I've never heard of anyone using a static JNI library; while it's
technically possible, it requires rebuilding and relinking the JVM.)